CSMR: Populate parse failures in source status.
Populate source status for all read / transform / validate steps in
the parser.  Includes fixes that should have been in the first
attempt.

Bug: 169158287

+++ e2e/testcases/status_monitoring.bats
@@ -3,6 +3,7 @@
 set -euo pipefail
 
 load "../lib/assert"
+load "../lib/env"
 load "../lib/git"
 load "../lib/setup"
 load "../lib/wait"
@@ -22,16 +23,16 @@ test_teardown() {
   setup::git::remove_all acme
 
   # Force delete repo object.
-  kubectl delete repo repo --ignore-not-found
-
+  if ! env::csmr; then
+    kubectl delete repo repo --ignore-not-found
+  fi
 }
 
 function ensure_error_free_repo () {
   debug::log "Ensure that the repo is error-free at the start of the test"
   git::commit -a -m "Commit the repo contents."
-
-  wait::for -t 30 -o "" -- kubectl get repo repo \
-    --output='jsonpath={.status.import.errors[0].errorMessage}'
+  wait::for nomos::repo_synced
+  nomos::import_error_code ""
 }
 
 @test "${FILE_NAME}: Invalid policydir gets an error message in status.source.errors" {
@@ -41,30 +42,38 @@ function ensure_error_free_repo () {
 
   kubectl apply -f "${MANIFEST_DIR}/importer_acme.yaml"
   nomos::restart_pods
-
-  wait::for -t 10 -o "" -- kubectl get repo repo \
-    --output='jsonpath={.status.import.errors[0].errorMessage}'
-  wait::for -t 10 -o "" -- kubectl get repo repo \
-    --output='jsonpath={.status.source.errors[0].errorMessage}'
+  nomos::import_error_code ""
+  nomos::source_error_code ""
 
   debug::log "Break the policydir in the repo"
-  kubectl apply -f "${MANIFEST_DIR}/importer_some-nonexistent-policydir.yaml"
-  nomos::restart_pods
+  if env::csmr; then
+    kubectl patch -n config-management-system rootsyncs.configsync.gke.io root-sync \
+      --type merge --patch '{"spec":{"git":{"dir":"some-nonexistent-policydir"}}}'
+  else
+    kubectl apply -f "${MANIFEST_DIR}/importer_some-nonexistent-policydir.yaml"
+    nomos::restart_pods
+  fi
 
   # Increased timeout from initial 30 to 60 for flakiness.  git-importer
   # gets restarted on each object change.
   debug::log "Expect an error to be present in status.source.errors"
-  wait::for -t 90 -c "some-nonexistent-policydir" -- \
-    kubectl get repo repo -o=yaml \
-      --output='jsonpath={.status.source.errors[0].errorMessage}'
+  if env::csmr; then
+    nomos::source_error_code "2001"
+  else
+    nomos::source_error_code "2004"
+  fi
 
   debug::log "Fix the policydir in the repo"
-  kubectl apply -f "${MANIFEST_DIR}/importer_acme.yaml"
-  nomos::restart_pods
+  if env::csmr; then
+    kubectl patch -n config-management-system rootsyncs.configsync.gke.io root-sync \
+      --type merge --patch '{"spec":{"git":{"dir":"acme"}}}'
+  else
+    kubectl apply -f "${MANIFEST_DIR}/importer_acme.yaml"
+    nomos::restart_pods
+  fi
 
   debug::log "Expect repo to recover from the error in source message"
-  wait::for -t 90 -o "" -- kubectl get repo repo \
-    --output='jsonpath={.status.source.errors[0].errorMessage}'
+  nomos::source_error_code ""
 }
 
 function namespace_count_matches () {
@@ -75,35 +84,34 @@ function namespace_count_matches () {
   [ "${expected}" = "${actual}" ]
 }
 
-@test "${FILE_NAME}: Deleting all namespaces gets an error message in status.source.errors" {
+function check_acme_ns() {
+  local -a ACME_NS
+  # shellcheck disable=SC2038
+  mapfile -t ACME_NS < <(find acme -name namespace.yaml | xargs dirname | sed 's|.*/||')
+  for ns in "${ACME_NS[@]}"; do
+    kubectl get ns "$ns"
+  done
+}
 
+@test "${FILE_NAME}: Deleting all namespaces gets an error message in status.source.errors" {
   cd "${TEST_REPO}"
   mkdir -p acme/namespaces
   cp -r "${NOMOS_DIR}/examples/acme/namespaces" acme
   git add -A
   ensure_error_free_repo
-
-  # the safety namespace (see lib/setup.bash) may be terminating, ignore it
-  SYSTEM_NS_COUNT=$(kubectl get ns --no-headers | grep -c -v safety)
-  ACME_NS_COUNT=$(find acme -name namespace.yaml | wc -l)
-  (( EXPECTED_NS_COUNT = SYSTEM_NS_COUNT + ACME_NS_COUNT ))
-  debug::log "Before delete: ensure we have the ${EXPECTED_NS_COUNT} namespaces we expect"
-  wait::for -t 30 -- namespace_count_matches "${EXPECTED_NS_COUNT}"
+  check_acme_ns
 
   debug::log "Delete all the namespaces (oops!)"
-
   cd "${TEST_REPO}"
   rm -rf "${TEST_REPO:?}/acme/namespaces"
   git add -A
   git::commit -a -m "wiping all namespaces from acme"
 
   debug::log "Expect an error to be present in status.source.errors"
-  wait::for -t 90 -c "mounted git repo appears to contain no managed namespaces" -- \
-    kubectl get repo repo -o=yaml \
-      --output='jsonpath={.status.source.errors[0].errorMessage}'
+  nomos::source_error_code "2006"
 
-  debug::log "Before restore: ensure we still have the ${EXPECTED_NS_COUNT} namespaces we expect"
-  wait::for -t 30 -- namespace_count_matches "${EXPECTED_NS_COUNT}"
+  debug::log "Before restore: ensure we still have the acme namespaces we expect"
+  check_acme_ns
 
   debug::log "Restore the namespaces"
   cd "${TEST_REPO}"
@@ -112,11 +120,10 @@ function namespace_count_matches () {
   git::commit -a -m "restoring the namespaces"
 
   debug::log "Expect repo to recover from the error in source message"
-  wait::for -t 90 -o "" -- kubectl get repo repo \
-    --output='jsonpath={.status.source.errors[0].errorMessage}'
+  nomos::source_error_code ""
 
-  debug::log "After restore: ensure we still have the ${EXPECTED_NS_COUNT} namespaces we expect"
-  wait::for -t 30 -- namespace_count_matches "${EXPECTED_NS_COUNT}"
+  debug::log "After restore: ensure we still have the acme namespaces we expect"
+  check_acme_ns
 
   setup::git::remove_all acme
 }
